requirements.txt
fastapi
uvicorn
python-dotenv
openrouter
requests
